Update Raspbian and Go
Update Go to 1.25.4. Also update Raspbian Lite to 2025-10-1. For some
reason, the date on the directory and on the image file is different, so
add raspbian_date variable so the build won't fail when these two values
are different. And while we're here, also replace debug provisioner with
breakpoint. These were added by mistake. Lastly, remove the password
change enforcement on the created user. This causes sometimes issues on
boot since the first boot process is actually multiple reboots in a row.